Changed 'ref' prop to 'reference'
It appears that 'ref' is special attribute and can't be a prop. So to pass ref to button on dynamically use "reference" prop. 

Thing to note. If invisible captcha element has 'ref' set to 'captcha' and 'reference' is set to 'reButton', then button can be accessed by `this.$refs.captcha.$refs.reButton`.